# job-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Plateforme Emploi",
        "description": "Documentation Swagger des endpoints Admin pour la gestion des utilisateurs, offres et candidatures",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/candidate/applications": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "Lister les candidatures du candidat connecté",
                "operationId": "aa73ca3f50b108d97caf221d009f8efc",
                "responses": {
                    "200": {
                        "description": "Liste paginée des candidatures du candidat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Application"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé - Candidat requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/jobs/{id}/apply": {
            "post": {
                "tags": [
                    "Applications"
                ],
                "summary": "Postuler à une offre d'emploi",
                "operationId": "810171efa09900eb416b44ee81c6f90d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du job à postuler",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ApplicationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Candidature soumise",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Application"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Offre non trouvée"
                    },
                    "422": {
                        "description": "Déjà postulé ou erreur de validation"
                    },
                    "403": {
                        "description": "Accès refusé - Candidat requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/candidate/applications/{id}": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "Afficher les détails d'une candidature spécifique du candidat",
                "operationId": "d27c4d6d9388e0a78abec8bb2f69b99a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la candidature",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la candidature",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Application"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Candidature non trouvée"
                    },
                    "403": {
                        "description": "Accès refusé - Candidat requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/employer/applications": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "Lister toutes les candidatures reçues par les offres de l'employeur connecté",
                "operationId": "107fcdddbae858d17fd4862a0ca26518",
                "responses": {
                    "200": {
                        "description": "Liste paginée des candidatures reçues",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Application"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé - Employeur requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/applications/{id}/status": {
            "put": {
                "tags": [
                    "Applications"
                ],
                "summary": "Mettre à jour le statut d'une candidature (Employeur uniquement)",
                "operationId": "9076be571c7443f1245225ded16b6a83",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la candidature",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ApplicationStatusRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statut mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Application"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Candidature non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Application": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "job_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "candidate_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "cover_letter": {
                        "type": "string",
                        "example": "Lettre de motivation..."
                    },
                    "resume": {
                        "type": "string",
                        "example": "cv.pdf"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "reviewed",
                            "accepted",
                            "rejected"
                        ],
                        "example": "pending"
                    },
                    "feedback": {
                        "type": "string",
                        "example": "Feedback de l'employeur"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "job": {
                        "$ref": "#/components/schemas/Job"
                    },
                    "candidate": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "ApplicationRequest": {
                "required": [
                    "cover_letter"
                ],
                "properties": {
                    "cover_letter": {
                        "type": "string",
                        "maxLength": 2000,
                        "minLength": 50,
                        "example": "Je suis très intéressé par cette opportunité..."
                    },
                    "resume": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "mon_cv.pdf"
                    }
                },
                "type": "object"
            },
            "ApplicationStatusRequest": {
                "required": [
                    "status"
                ],
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "reviewed",
                            "accepted",
                            "rejected"
                        ],
                        "example": "reviewed"
                    },
                    "feedback": {
                        "type": "string",
                        "maxLength": 1000,
                        "example": "Votre profil a été retenu pour un entretien"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User model",
                "required": [
                    "id",
                    "name",
                    "email",
                    "type"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "type": {
                        "type": "string",
                        "example": "candidate"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+33123456789"
                    },
                    "bio": {
                        "type": "string",
                        "example": "Développeur fullstack"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-22T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-22T10:00:00Z"
                    }
                },
                "type": "object"
            },
...
            "Job": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "employer_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Développeur Laravel Senior"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description du poste..."
                    },
                    "company": {
                        "type": "string",
                        "example": "Tech Solutions"
                    },
                    "location": {
                        "type": "string",
                        "example": "Paris, France"
                    },
                    "salary": {
                        "type": "number",
                        "format": "float",
                        "example": 55000
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "full_time",
                            "part_time",
                            "contract",
                            "internship"
                        ],
                        "example": "full_time"
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "technology",
                            "healthcare",
                            "education",
                            "finance",
                            "other"
                        ],
                        "example": "technology"
                    },
                    "application_deadline": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-31"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}